refactor: split over-budget files by responsibility for the derived settings form
Move the catalog snapshot/preset/reread validators into
gui-s3-catalog-validation, the flat settings action values into the
kernel settings-action-values module, the settings read result type and
its construction into daemon-settings-read-types, and the repository
settings panel out of SettingsView so every touched file fits its
merge-base line budget without shaving.
