+++ packages/kernel/src/domain/index.ts
@@ -266,7 +266,6 @@ export {
   parseLocalSettings,
   readSettingsFacet,
   repositorySettings,
-  repositorySettingsActionValues,
   serializeLocalSettings,
   validateRepositorySettings,
   validateSettingsV1,
@@ -287,6 +286,7 @@ export { decodeVerticalDefinition } from "../schemas/vertical-definition.ts";
 export type { VerticalDefinition } from "../schemas/vertical-definition.ts";
 export { buildVerticalDeclarationRead, validateVerticalDeclarationRead } from "./vertical-declaration.ts";
 export { settingsActionLocale, settingsUpdateInputFields } from "./settings-action-contract.ts";
+export { repositorySettingsActionValues } from "./settings-action-values.ts";
 export type { SettingsActionDraft } from "./settings-action-contract.ts";
 export {
   applyPeopleRosterAction,
