+++ packages/kernel/test/contracts/settings-action.test.ts
@@ -3,7 +3,7 @@ import assert from "node:assert/strict";
 import test from "node:test";
 import { explainEntityKind, getExecutableEntityAction } from "../../src/domain/entity-kind-registry.ts";
 import { SettingsActionError, settingsUpdateInputFields } from "../../src/domain/settings-action-contract.ts";
-import { repositorySettingsActionValues } from "../../src/domain/settings.ts";
+import { repositorySettingsActionValues } from "../../src/domain/settings-action-values.ts";
 import { assertSettingsEventInputs } from "../../src/domain/settings-event.ts";
 import { effectiveCloseoutGates } from "../../src/domain/settings-closeout.ts";
 import { readSettingsFacet, repositorySettings } from "../../src/domain/settings.ts";
